Make this more sane.
Stuffing mutable data into the controller instance is going to cause massive lossage,
as in a pre-fork environment (i.e. any non-test env) then you're likely to have different processes
get subsequent requests - meaning that requests will see each-other's tokens, and so break.

Instead, pass the ctx into the controller methods as a parameter, as this means we keep all the state
in the context where it should be, which makes this work correctly